+++ Modules/DataModel/Sources/PocketCastsDataModel/Private/Managers/Util/DatabaseHelper.swift
@@ -703,6 +703,17 @@ class DatabaseHelper {
             }
         }
 
+        if schemaVersion < 47 {
+            do {
+                try db.executeUpdate("ALTER TABLE SJEpisode ADD COLUMN contentType TEXT;", values: nil)
+                try db.executeUpdate("ALTER TABLE SJUserEpisode ADD COLUMN contentType TEXT;", values: nil)
+                schemaVersion = 47
+            } catch {
+                failedAt(47)
+                return
+            }
+        }
+
         db.commit()
     }
 }