+++ podcasts/PlaybackManager.swift
@@ -23,6 +23,7 @@ class PlaybackManager: ServerPlaybackDelegate {
         didSet {
             if sleepOnEpisodeEnd {
                 sleepTimeRemaining = -1
+                sleepTimerManager.recordSleepTimerDuration(duration: nil, onEpisodeEnd: true)
             }
             NotificationCenter.postOnMainThread(notification: Constants.Notifications.sleepTimerChanged)
         }
@@ -59,6 +60,8 @@ class PlaybackManager: ServerPlaybackDelegate {
     lazy var bookmarkManager: BookmarkManager = {
         BookmarkManager(playbackManager: self)
     }()
+
+    private lazy var sleepTimerManager = SleepTimerManager()
 
     /// The player we should fallback to
     private var fallbackToPlayer: PlaybackProtocol.Type? = nil
@@ -236,6 +239,8 @@ class PlaybackManager: ServerPlaybackDelegate {
             }
 
             self.updateIdleTimer()
+
+            self.sleepTimerManager.restartSleepTimerIfNeeded()
         })
     }
 
@@ -945,7 +950,7 @@ class PlaybackManager: ServerPlaybackDelegate {
 
     func playerDidFinishPlayingEpisode() {
         if sleepOnEpisodeEnd {
-            pause()
+            pauseAndRecordSleepTimerFinished()
             cancelSleepTimer()
             return
         }
@@ -1363,16 +1368,25 @@ class PlaybackManager: ServerPlaybackDelegate {
 
         // here (as above) we're assuming that in general the timer fires around once a second. Might have to investigate this though as it might not always be the case
         if sleepTimeRemaining >= 0 {
+            if sleepTimeRemaining == sleepTimerManager.sleepTimerFadeDuration {
+                sleepTimerManager.performFadeOut(player: player)
+            }
+
             sleepTimeRemaining = sleepTimeRemaining - updateTimerInterval
 
             if sleepTimeRemaining < 0 {
-                pause()
+                pauseAndRecordSleepTimerFinished()
             }
         }
 
         if player.buffering() == false {
             updateChapterInfo()
         }
+    }
+
+    private func pauseAndRecordSleepTimerFinished() {
+        sleepTimerManager.recordSleepTimerFinished()
+        pause()
     }
 
     private func fireProgressNotification() {
@@ -1462,7 +1476,8 @@ class PlaybackManager: ServerPlaybackDelegate {
 
     // MARK: - Sleep Timer
 
-    func cancelSleepTimer() {
+    func cancelSleepTimer(userInitiated: Bool = false) {
+        sleepTimerManager.cancelSleepTimer(userInitiated: userInitiated)
         sleepTimeRemaining = -1
         sleepOnEpisodeEnd = false
         NotificationCenter.postOnMainThread(notification: Constants.Notifications.sleepTimerChanged)
@@ -1473,9 +1488,21 @@ class PlaybackManager: ServerPlaybackDelegate {
     }
 
     func setSleepTimerInterval(_ stopIn: TimeInterval) {
+        sleepTimerManager.recordSleepTimerDuration(duration: stopIn, onEpisodeEnd: nil)
         sleepTimeRemaining = stopIn
         NotificationCenter.postOnMainThread(notification: Constants.Notifications.sleepTimerChanged)
         Analytics.track(.playerSleepTimerEnabled, properties: ["time": Int(stopIn)])
+    }
+
+    func restartSleepTimer() {
+        guard sleepTimerActive() else {
+            return
+        }
+
+        #if !os(watchOS)
+        Toast.show(L10n.deviceShakeSleepTimer)
+        #endif
+        sleepTimerManager.restartSleepTimer()
     }
 
     // MARK: - Remote Control support