+++ CHANGELOG.md
@@ -1,6 +1,14 @@
+7.63
+-----
+- Show the Up Next bar on the tab bar. [#1613]
+- New design for the Podcasts grid layouts. [#1628]
+- Shake the device to restart Sleep Timer [#1627]
+- Playback fades out when finishing a sleep timer [#1629]
+
 7.62
 -----
-
+- Sleep Timer restarts automatically if you play again within 5 minutes [#1612]
+- Add redesigned Categories picker to the top of Discover. [#1621]
 
 7.61
 -----